+++ src/pools/selection-strategies/interleaved-weighted-round-robin-worker-choice-strategy.ts
@@ -96,9 +96,6 @@ export class InterleavedWeightedRoundRobinWorkerChoiceStrategy<
         workerNodeKey++
       ) {
         this.workerNodeId = workerNodeKey
-        if (!this.isWorkerNodeEligible(workerNodeKey)) {
-          continue
-        }
         if (
           this.workerNodeId !== this.nextWorkerNodeKey &&
           this.workerVirtualTaskRunTime !== 0