+++ src/pools/selection-strategies/least-busy-worker-choice-strategy.ts
@@ -74,9 +74,6 @@ export class LeastBusyWorkerChoiceStrategy<
     let minTime = Infinity
     let chosenWorkerNodeKey: number | undefined
     for (const [workerNodeKey, workerNode] of this.pool.workerNodes.entries()) {
-      if (!this.isWorkerNodeEligible(workerNodeKey)) {
-        continue
-      }
       const workerTime =
         (workerNode.usage.runTime?.aggregate ?? 0) +
         (workerNode.usage.waitTime?.aggregate ?? 0)