+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Fixed
+
+- Fix race condition in worker choice strategies at worker node info querying while not yet initialized .
+
 ## [2.6.38] - 2023-08-30
 
 ### Added